bin/Linux/adb-1.0.32: add new version of adb bin/Linux/after-check-out.sh: add ja_JP bin/Linux/config/pkgs/shadowsocks: add my vpn app bin/Linux/config/upd_system: no need to test is-debian bin/adb-is-activity: fix adb-install screwing over my clipboard bin/after-co-ln-s.sh: should not back up when rc.local not exist (on new debian versions) bin/ask-for-input-with-sawfish: fix for getopts bin/download-google-sdk.sh: should work for when the file download is incomplete, wget -N won't cover this bin/mm-ant: should not print the passwd bin/mm: fix for 'build is ok' always printed bin/overide/repo: should not display repo not found bin/start-ap: try to make it work for my usb wifi bin/start-http-proxy: should change the port, 8888 is the system daemon bin/stop-ap: fix for usb wifi bin/sync-new-pc: should sync files in the more desirable order bin/system-code-reading: do code reading index for system directorys bin/update-android-mirror: 8888 is system tinyproxy's port, I can't kill it without sudo

+++ bin/after-co-ln-s.sh
@@ -329,7 +329,7 @@ if ask-if-not-bhj "Do you want to switch the ctrl/alt, esc/caps_lock keys?"; the
 fi
 
 if test -L /etc/rc.local || can-sudo-and-yes-or-no-p "Replace /etc/rc.local with system-config's version?"; then
-    if ! test -L /etc/rc.local; then
+    if test -e /etc/rc.local && ! test -L /etc/rc.local; then
         sudo cp /etc/rc.local /etc/rc.local.bak
     fi
     if test "$(readlink -f /etc/rc.local)" != ~/system-config/etc/rc.local; then
